+++ app/javascript/vue/tasks/biological_associations/new/app.vue
@@ -78,18 +78,24 @@ defineOptions({
   name: 'NewBiologicalAssociation'
 })
 
+const store = useStore()
+
 const shortcuts = ref([
   {
     keys: [platformKey(), 's'],
     handler() {
       store.saveBiologicalAssociation()
     }
+  },
+  {
+    keys: [platformKey(), 'l'],
+    handler() {
+      store.toggleLock()
+    }
   }
 ])
 
 useHotkey(shortcuts.value)
-
-const store = useStore()
 
 const currentBiologicalAssociation = computed(() =>
   store.biologicalAssociations.find(
@@ -105,12 +111,24 @@ onBeforeMount(() => {
     'Save and create new biological association',
     'New biological association'
   )
+
+  TW.workbench.keyboard.createLegend(
+    `${platformKey()}+s`,
+    'Lock all',
+    'New biological association'
+  )
 })
 </script>
 
 <style scoped>
 .grid-panels {
-  display: grid;
-  grid-template-columns: 1fr 1fr 0.75fr;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+.panel {
+  flex: 1 1 20%;
+  box-sizing: border-box;
 }
 </style>