+++ app/javascript/vue/tasks/digitize/components/collectionObject/Citation/CitationMain.vue
@@ -11,11 +11,8 @@
       @source="addLabel"
       @submit="addCitation"
     >
-      <template #smart-selector-right>
-        <v-lock
-          class="margin-small-left"
-          v-model="lock.coCitations"
-        />
+      <template #tabs-right>
+        <VLock v-model="lock.coCitations" />
       </template>
     </FormCitation>
     <display-list