+++ app/models/otu.rb
@@ -46,6 +46,7 @@ class Otu < ApplicationRecord
   include Otu::MatrixHooks
   include Otu::Maps
 
+  include Shared::DwcOccurrenceHooks
   include Shared::IsData
 
   include Shared::QueryBatchUpdate
@@ -521,6 +522,16 @@ class Otu < ApplicationRecord
         end
       end
     end
+  end
+
+  def dwc_occurrences
+    a = ::Queries::DwcOccurrence::Filter.new( asserted_distribution_query: {otu_id: id, project_id:},).all
+    b = ::Queries::DwcOccurrence::Filter.new( collection_object_query: {otu_id: id, project_id:},).all
+    # TODO FieldOccurrence in same pattern
+
+    ::Queries.union(
+      ::DwcOccurrence, [ a, b ]
+    )
   end
 
   protected