+++ CHANGELOG.md
@@ -9,22 +9,38 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 
+- Created/updated overviews for user data
 - Unify objects task [#970]
 - Attribution to ObservationMatrix
 - New biological association task [#4026]
+- Duplicate OTU predictor task [#4083]
+- DwcOccurrenceHooks for BiocurationGroups, OTUs
+- New image task: is original checkbox to source panel [#4090]
+- Confidence facets and batch operations to all Filters [#4043]
 
 ### Changed
 
+- RecordNumber identifiers can be duplicated (namespace + identifier combinations) across CollectionObjects [#4096]
 - Local identifier display in CollectionObject tag now prefers position to break tie with RecordNumber and CatalogNumber [#4074]
+- Filter CO: Show only current taxon determination [#4092]
 
 ### Fixed
 
+- Simple TaxonName batch load failing with invalid children
+- `api/v1/images/:id` broken for non-integer ids
+- Containerizing objects prevented identific increments
 - New dichotomous key: Radial annotator is not loading the correct data [#4076]
 
-[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/4074
 [#4026]: https://github.com/SpeciesFileGroup/taxonworks/issues/4026
+[#4043]: https://github.com/SpeciesFileGroup/taxonworks/issues/4043
 [#4074]: https://github.com/SpeciesFileGroup/taxonworks/issues/4074
 [#4076]: https://github.com/SpeciesFileGroup/taxonworks/issues/4076
+[#4077]: https://github.com/SpeciesFileGroup/taxonworks/issues/4077
+[#4083]: https://github.com/SpeciesFileGroup/taxonworks/issues/4083
+[#4090]: https://github.com/SpeciesFileGroup/taxonworks/issues/4090
+[#4092]: https://github.com/SpeciesFileGroup/taxonworks/issues/4092
+[#4096]: https://github.com/SpeciesFileGroup/taxonworks/issues/4096
+[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
 
 ## [0.44.3] - 2024-10-03